Downloads/ut-wireless-sim-SRTupdated/ut-wireless-sim/backend/requirements.txt
fastapi
uvicorn[standard]
sionna-rt
numpy
scipy
requests
python-dotenv
